move logic into src, export it from lib
